# docs/swagger.yaml
basePath: /
definitions:
  controllers.CreateEventRequest:
    properties:
      name:
        type: string
      slug:
        type: string
    type: object
  controllers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  controllers.SignUpRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        description: 'optional: "admin" or "attendee" (defaults to "attendee")'
        type: string
    type: object
  helpers.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  helpers.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/helpers.APIError'
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the backend API for a conference ticketing system.
  title: Multi-Track Ticketing API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password. Returns a JWT containing
        user id, email, and roles.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: data contains token and token_type
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: 'error.code: bad_request'
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "401":
          description: 'error.code: unauthorized'
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "500":
          description: 'error.code: internal_error'
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      summary: Log in
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: 'Create a new user with email, password, and name. Optional role:
        "admin" or "attendee" (defaults to "attendee"). Password is stored hashed.'
      parameters:
      - description: Sign-up data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: data contains the created user
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: 'error.code: bad_request'
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "500":
          description: 'error.code: internal_error'
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      summary: Sign up a new user
      tags:
      - auth
  /events:
    post:
      consumes:
      - application/json
      description: Create a new conference event. Only name and slug are accepted
        in the body; id and timestamps are server-generated.
      parameters:
      - description: Event data (name and slug only)
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: data contains the created event
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: 'error.code: bad_request'
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "500":
          description: 'error.code: internal_error'
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      summary: Create a new event
      tags:
      - events
  /events/{eventID}/import/sessionize/{sessionizeID}:
    post:
      description: Import rooms and sessions from Sessionize for a specific event
      parameters:
      - description: Event ID
        in: path
        name: eventID
        required: true
        type: string
      - description: Sessionize ID
        in: path
        name: sessionizeID
        required: true
        type: string
      responses:
        "200":
          description: data contains status message
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "400":
          description: 'error.code: bad_request'
          schema:
            $ref: '#/definitions/helpers.APIResponse'
        "500":
          description: 'error.code: internal_error'
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      summary: Import schedule from Sessionize
      tags:
      - events
swagger: "2.0"